+++ code/src/sixsq/nuvla/server/resources/module.clj
@@ -9,7 +9,6 @@ component, or application.
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.db.impl :as db]
-    [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
@@ -113,46 +112,6 @@ component, or application.
         crud/validate)
     {}))
 
-
-(defn active-claim->account-id
-  [active-claim]
-  (let [filter     (format "parent='%s'" active-claim)
-        options    {:cimi-params {:filter (parser/parse-cimi-filter filter)}}
-        account-id (-> (crud/query-as-admin "vendor" options)
-                       second
-                       first
-                       :account-id)]
-    (or account-id
-        (throw (r/ex-response (str "unable to resolve vendor account-id for active-claim '"
-                                   active-claim "' ") 409)))))
-
-
-(defn set-price
-  [{{:keys [price-id cent-amount-daily currency follow-customer-trial] :as price}
-    :price name :name path :path :as body}
-   active-claim]
-  (if price
-    (let [product-id (some-> price-id pricing-impl/retrieve-price
-                             pricing-impl/price->map :product-id)
-          account-id (active-claim->account-id active-claim)
-          s-price    (pricing-impl/create-price
-                       (cond-> {"currency"    currency
-                                "unit_amount" cent-amount-daily
-                                "recurring"   {"interval"        "month"
-                                               "aggregate_usage" "sum"
-                                               "usage_type"      "metered"}}
-                               product-id (assoc "product" product-id)
-                               (nil? product-id) (assoc "product_data" {"name"       (or name path)
-                                                                        "unit_label" "day"})))
-          price      (cond-> {:price-id          (pricing-impl/get-id s-price)
-                              :product-id        (pricing-impl/get-product s-price)
-                              :account-id        account-id
-                              :cent-amount-daily cent-amount-daily
-                              :currency          currency}
-                             (some? follow-customer-trial) (assoc :follow-customer-trial follow-customer-trial))]
-      (assoc body :price price))
-    body))
-
 (defn throw-cannot-access-registries-or-creds
   [{{{:keys [private-registries registries-credentials]} :content} :body :as request}]
   (when
@@ -224,7 +183,7 @@ component, or application.
             (assoc :versions [(cond-> {:href   content-id
                                        :author author}
                                       commit (assoc :commit commit))])
-            (set-price (auth/current-active-claim request))
+            (utils/set-price (auth/current-active-claim request))
             (db-add-module-meta request))))))
 
 
@@ -242,6 +201,7 @@ component, or application.
           is-not-project? (not (utils/is-project? subtype))]
       (-> module-meta
           (cond-> is-not-project? (utils/get-module-content uuid))
+          utils/resolve-vendor-email
           (crud/set-operations request)
           (a/select-viewable-keys request)
           (r/json-response)))
@@ -316,7 +276,7 @@ component, or application.
               (assoc :body
                      (cond-> module-meta
                              price-changed? (-> (assoc :price (merge price (:price module-meta)))
-                                                (set-price (auth/current-active-claim request)))
+                                                (utils/set-price (auth/current-active-claim request)))
                              versions (assoc :versions versions)))
               edit-impl))))
     (catch Exception e