+++ code/src/sixsq/nuvla/server/resources/module/utils.clj
@@ -2,6 +2,8 @@
   (:require
     [clojure.edn :as edn]
     [clojure.string :as str]
+    [sixsq.nuvla.db.filter.parser :as parser]
+    [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.util.response :as r]))
 
@@ -67,8 +69,8 @@
   (loop [i (dec (count versions))]
     (when (not (neg? i))
       (if (some? (nth versions i))
-       i
-       (recur (dec i))))))
+        i
+        (recur (dec i))))))
 
 
 (defn retrieve-content-id
@@ -96,3 +98,54 @@
                            (throw (r/ex-not-found
                                     (str "Module version not found: " id)))))]
     (assoc module-meta :content module-content)))
+
+(defn get-vendor-by-query-as-admin
+  [filter-str]
+  (let [options {:cimi-params {:filter (parser/parse-cimi-filter filter-str)}}]
+    (-> (crud/query-as-admin "vendor" options)
+        second
+        first)))
+
+(defn active-claim->account-id
+  [active-claim]
+  (or (some->> active-claim
+               (format "parent='%s'")
+               get-vendor-by-query-as-admin
+               :account-id)
+      (throw (r/ex-response (str "unable to resolve vendor account-id for active-claim '"
+                                 active-claim "' ") 409))))
+
+(defn resolve-vendor-email
+  [{{:keys [account-id]} :price :as module-meta}]
+  (if-let [email (some->> account-id
+                          (format "account-id='%s'")
+                          get-vendor-by-query-as-admin
+                          :email)]
+    (assoc-in module-meta [:price :vendor-email] email)
+    module-meta))
+
+(defn set-price
+  [{{:keys [price-id cent-amount-daily currency follow-customer-trial] :as price}
+    :price name :name path :path :as body}
+   active-claim]
+  (if price
+    (let [product-id (some-> price-id pricing-impl/retrieve-price
+                             pricing-impl/price->map :product-id)
+          account-id (active-claim->account-id active-claim)
+          s-price    (pricing-impl/create-price
+                       (cond-> {"currency"    currency
+                                "unit_amount" cent-amount-daily
+                                "recurring"   {"interval"        "month"
+                                               "aggregate_usage" "sum"
+                                               "usage_type"      "metered"}}
+                               product-id (assoc "product" product-id)
+                               (nil? product-id) (assoc "product_data" {"name"       (or name path)
+                                                                        "unit_label" "day"})))
+          price      (cond-> {:price-id          (pricing-impl/get-id s-price)
+                              :product-id        (pricing-impl/get-product s-price)
+                              :account-id        account-id
+                              :cent-amount-daily cent-amount-daily
+                              :currency          currency}
+                             (some? follow-customer-trial) (assoc :follow-customer-trial follow-customer-trial))]
+      (assoc body :price price))
+    body))